+++ src/design/plone/contenttypes/behaviors/dataset_correlati.py
@@ -30,8 +30,8 @@ class IDatasetCorrelati(model.Schema):
         RelatedItemsFieldWidget,
         vocabulary="plone.app.vocabularies.Catalog",
         pattern_options={
-            "recentlyUsed": True,  # Just turn on. Config in plone.app.widgets.
             "selectableTypes": ["Dataset"],
+            "maximumSelectionSize": 50,
         },
     )